webapp/requirements.txt
Django<1.9
django-tastypie
uwsgi
pycrypto
requests
sendgrid
